[108] normal mode 기능 구현 완료
